requirements.txt
python-dotenv==1.0.0
requests==2.31.0
facebook-sdk==3.1.0
calendly==1.1.1
pandas==2.1.0
scikit-learn==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose==3.3.0
aiohttp==3.11.14
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
fastapi-cache2==0.2.1
prometheus-fastapi-instrumentator==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
tenacity==8.2.3
structlog==23.2.0 